move files from fdr_analysis to fdrx module
